git-svn-id: svn://192.168.1.132/autoStock/autoStock@709 33c8eb2e-3fb0-4f7a-9790-3894f54e5fb4

+++ src/com/autoStock/signal/SignalOfEncog.java
@@ -49,7 +49,7 @@ public class SignalOfEncog extends SignalBase {
 		
 		MLData input = new BasicMLData(basicNetwork.getInputCount());
 		
-		int[] inputWindow = encogInputWindow.getAsWindow();
+		double[] inputWindow = encogInputWindow.getAsWindow();
 		
 		if (inputWindow.length != basicNetwork.getInputCount()){
 			throw new IllegalArgumentException("Input sizes don't match, supplied, needed: " + inputWindow.length + ", " + basicNetwork.getInputCount());
